# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * RazeBot website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStats {
  servers: number;
  users: number;
  commandsRun: number;
  uptime: string;
}

export interface BotCommand {
  name: string;
  description: string;
  /** @nullable */
  usage?: string | null;
}

export interface CommandCategory {
  category: string;
  icon: string;
  commands: BotCommand[];
}

export type AnnouncementType = typeof AnnouncementType[keyof typeof AnnouncementType];


export const AnnouncementType = {
  info: 'info',
  warning: 'warning',
  update: 'update',
  maintenance: 'maintenance',
} as const;

export interface Announcement {
  id: string;
  title: string;
  content: string;
  type: AnnouncementType;
  createdAt: string;
}

export type AnnouncementInputType = typeof AnnouncementInputType[keyof typeof AnnouncementInputType];


export const AnnouncementInputType = {
  info: 'info',
  warning: 'warning',
  update: 'update',
  maintenance: 'maintenance',
} as const;

export interface AnnouncementInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  type: AnnouncementInputType;
}

export interface Partner {
  id: string;
  name: string;
  description: string;
  inviteUrl: string;
  /** @nullable */
  iconUrl?: string | null;
  /** @nullable */
  memberCount?: string | null;
  createdAt: string;
}

export interface PartnerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  inviteUrl: string;
  /** @nullable */
  iconUrl?: string | null;
  /** @nullable */
  memberCount?: string | null;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  token: string;
}

export interface AdminStats {
  servers: number;
  users: number;
  commandsRun: number;
  uptime: string;
  announcements: number;
  partners: number;
  lastUpdated: string;
}

export interface StatsOverrideInput {
  /** @nullable */
  servers?: number | null;
  /** @nullable */
  users?: number | null;
  /** @nullable */
  commandsRun?: number | null;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}