# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RazeBot website API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stats
    description: Bot statistics
  - name: commands
    description: Bot commands
  - name: announcements
    description: Announcements
  - name: partners
    description: Partners
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /stats:
    get:
      operationId: getBotStats
      tags: [stats]
      summary: Get bot statistics
      responses:
        "200":
          description: Bot statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStats"

  /commands:
    get:
      operationId: getCommands
      tags: [commands]
      summary: Get all bot commands grouped by category
      responses:
        "200":
          description: Commands list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommandCategory"

  /announcements:
    get:
      operationId: getAnnouncements
      tags: [announcements]
      summary: Get public announcements
      responses:
        "200":
          description: Announcements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"

  /partners:
    get:
      operationId: getPartners
      tags: [partners]
      summary: Get public partners list
      responses:
        "200":
          description: Partners list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get live admin statistics
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/announcements:
    post:
      operationId: createAnnouncement
      tags: [admin]
      summary: Create a new announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [admin]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/partners:
    post:
      operationId: createPartner
      tags: [admin]
      summary: Add a new partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/partners/{id}:
    delete:
      operationId: deletePartner
      tags: [admin]
      summary: Remove a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats-override:
    post:
      operationId: updateStatsOverride
      tags: [admin]
      summary: Override displayed bot stats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatsOverrideInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    BotStats:
      type: object
      required: [servers, users, commandsRun, uptime]
      properties:
        servers:
          type: integer
        users:
          type: integer
        commandsRun:
          type: integer
        uptime:
          type: string

    CommandCategory:
      type: object
      required: [category, icon, commands]
      properties:
        category:
          type: string
        icon:
          type: string
        commands:
          type: array
          items:
            $ref: "#/components/schemas/BotCommand"

    BotCommand:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        usage:
          type: ["string", "null"]

    Announcement:
      type: object
      required: [id, title, content, type, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [info, warning, update, maintenance]
        createdAt:
          type: string

    AnnouncementInput:
      type: object
      required: [title, content, type]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        type:
          type: string
          enum: [info, warning, update, maintenance]

    Partner:
      type: object
      required: [id, name, description, inviteUrl, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        inviteUrl:
          type: string
        iconUrl:
          type: ["string", "null"]
        memberCount:
          type: ["string", "null"]
        createdAt:
          type: string

    PartnerInput:
      type: object
      required: [name, description, inviteUrl]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        inviteUrl:
          type: string
          minLength: 1
        iconUrl:
          type: ["string", "null"]
        memberCount:
          type: ["string", "null"]

    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AdminLoginResult:
      type: object
      required: [success, token]
      properties:
        success:
          type: boolean
        token:
          type: string

    AdminStats:
      type: object
      required: [servers, users, commandsRun, uptime, announcements, partners, lastUpdated]
      properties:
        servers:
          type: integer
        users:
          type: integer
        commandsRun:
          type: integer
        uptime:
          type: string
        announcements:
          type: integer
        partners:
          type: integer
        lastUpdated:
          type: string

    StatsOverrideInput:
      type: object
      properties:
        servers:
          type: ["integer", "null"]
        users:
          type: ["integer", "null"]
        commandsRun:
          type: ["integer", "null"]

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean