# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * RazeBot website API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementType } from './announcementType';

export interface Announcement {
  id: string;
  title: string;
  content: string;
  type: AnnouncementType;
  createdAt: string;
}