# lib/api-zod/src/generated/types/partner.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * RazeBot website API
 * OpenAPI spec version: 0.1.0
 */

export interface Partner {
  id: string;
  name: string;
  description: string;
  inviteUrl: string;
  /** @nullable */
  iconUrl?: string | null;
  /** @nullable */
  memberCount?: string | null;
  createdAt: string;
}